requirements.txt
fastapi==0.111.0
mysql-connector-python==8.4.0
SQLAlchemy==2.0.31
pytest==8.2.2
